+++ src/main/java/com/plivo/api/models/base/Deleter.java
@@ -1,6 +1,7 @@
 package com.plivo.api.models.base;
 
 import com.plivo.api.PlivoClient;
+import com.plivo.api.exceptions.PlivoValidationException;
 import com.plivo.api.exceptions.PlivoRestException;
 import java.io.IOException;
 import okhttp3.ResponseBody;
@@ -15,6 +16,7 @@ import retrofit2.Response;
 public abstract class Deleter<T extends BaseResource> extends BaseRequest<T> {
 
   protected String id;
+  protected String secondaryId;
 
   public Deleter(String id) {
     this.id = id;
@@ -24,10 +26,18 @@ public abstract class Deleter<T extends BaseResource> extends BaseRequest<T> {
     }
   }
 
+  public Deleter(String id, String secondaryId) {
+    if (id == null || secondaryId == null) {
+      throw new IllegalArgumentException("id/secondaryId cannot be null");
+    }
+    this.id = id;
+    this.secondaryId = secondaryId;
+  }
+
   /**
    * Actually delete the resource.
    */
-  public void delete() throws IOException, PlivoRestException {
+  public void delete() throws IOException, PlivoRestException, PlivoValidationException {
     validate();
     Response<ResponseBody> response = obtainCall().execute();
 
@@ -41,5 +51,5 @@ public abstract class Deleter<T extends BaseResource> extends BaseRequest<T> {
   }
 
 
-  protected abstract Call<ResponseBody> obtainCall();
+  protected abstract Call<ResponseBody> obtainCall() throws PlivoValidationException;
 }