+++ src/main/java/com/plivo/api/xml/JaxbCharacterEscapeHandler.java
@@ -3,7 +3,7 @@ package com.plivo.api.xml;
 import java.io.IOException;
 import java.io.Writer;
 
-import com.sun.xml.internal.bind.marshaller.CharacterEscapeHandler;
+import com.sun.xml.bind.marshaller.CharacterEscapeHandler;
 
 public class JaxbCharacterEscapeHandler implements CharacterEscapeHandler {